build: Updating composer dependencies
* mediawiki/minus-x: 0.3.2 → 1.0.0
* mediawiki/mediawiki-phan-config: 0.9.0 → 0.9.1

+++ includes/CalendarTable.php
@@ -375,7 +375,7 @@ class CalendarTable {
 
 		if ( $offset != 0 ) {
 			$offM = $offset % 12;
-			$offset = round( ( $offset - $offM ) / 12 );
+			$offset = (int)round( ( $offset - $offM ) / 12 );
 			$this->month += $offM;
 			$this->year += $offset;
 			if ( $this->month > 12 ) {
